refactor: autoplay と intent の実装とテストをディレクトリ単位にまとめる
