+++ packages/react/src/components/Carousel/autoplay/useHoverPause.test.tsx
