+++ packages/react/src/components/Carousel/autoplay/useHoverPause.ts
