+++ packages/react/src/components/Carousel/index.test.tsx
@@ -1,80 +1,18 @@
 import { createRef } from 'react'
 import { render, screen, fireEvent, act } from '@testing-library/react'
 import { renderToString } from 'react-dom/server'
-import {
-  vi,
-  describe,
-  it,
-  expect,
-  beforeEach,
-  afterEach,
-  beforeAll,
-} from 'vitest'
+import { vi, describe, it, expect, beforeEach, afterEach } from 'vitest'
 import Carousel, {
   type CarouselHandlerRef,
   type CarouselDefaultScroll,
-  type CarouselProps,
 } from '.'
-
-// jsdom は :hover のライブ判定を持たず、matches(':hover') が実際の状態に関わらず
-// 常に true を返す。useHoverPause の種付けがテスト全体を止めないよう、
-// :hover の結果だけをテストから制御する（既定は「乗っていない」＝ false）。
-let stubbedHover = false
-beforeAll(() => {
-  const originalMatches = Element.prototype.matches
-  vi.spyOn(Element.prototype, 'matches').mockImplementation(function (
-    this: Element,
-    selector: string,
-  ) {
-    return selector === ':hover'
-      ? stubbedHover
-      : originalMatches.call(this, selector)
-  })
-})
-afterEach(() => {
-  stubbedHover = false
-})
+import { getScroller, mockScrollerGeometry } from './testUtils'
 
 const slides = Array.from({ length: 6 }, (_, i) => (
   <div key={`item-${i}`} data-testid={`slide-${i}`}>
     Slide {i}
   </div>
 ))
-
-function mockScrollerGeometry(
-  el: HTMLElement,
-  { scrollLeft = 0, scrollWidth = 2400, clientWidth = 800 } = {},
-) {
-  Object.defineProperty(el, 'scrollLeft', {
-    get: () => scrollLeft,
-    set: vi.fn(),
-    configurable: true,
-  })
-  Object.defineProperty(el, 'scrollWidth', {
-    value: scrollWidth,
-    configurable: true,
-  })
-  Object.defineProperty(el, 'clientWidth', {
-    value: clientWidth,
-    configurable: true,
-  })
-
-  for (let i = 0; i < el.children.length; i++) {
-    const child = el.children[i] as HTMLElement
-    Object.defineProperty(child, 'offsetLeft', {
-      value: i * 400,
-      configurable: true,
-    })
-    Object.defineProperty(child, 'offsetWidth', {
-      value: 380,
-      configurable: true,
-    })
-  }
-}
-
-function getScroller() {
-  return document.querySelector('.charcoal-carousel__scroller') as HTMLElement
-}
 
 // 各 observer が「observe した要素 → コールバック」を覚え、テストから任意要素の
 // intersection を発火できるようにする（新設計は item ごとに observer を持つ）。
@@ -1299,740 +1237,5 @@ describe('Carousel', () => {
       expect(html).not.toContain('data-clone')
       expect(html).toContain('data-testid="slide-0"')
     })
-  })
-})
-
-describe('onChange', () => {
-  // 中央検出を手動で駆動するため IntersectionObserver を差し替える
-  let triggerCenter: (el: Element) => void
-  let origIO: typeof globalThis.IntersectionObserver
-
-  beforeEach(() => {
-    // jsdom は onscrollend を持つが実イベントは発火しないため、debounce(100ms) 経路を
-    // 強制する（scrollSettle.ts の分岐は 'onscrollend' in window で判定するため）。
-    Reflect.deleteProperty(window, 'onscrollend')
-    const callbacks = new Map<Element, IntersectionObserverCallback>()
-    origIO = globalThis.IntersectionObserver
-    globalThis.IntersectionObserver = class {
-      constructor(private cb: IntersectionObserverCallback) {}
-      observe(el: Element) {
-        callbacks.set(el, this.cb)
-      }
-      unobserve(el: Element) {
-        callbacks.delete(el)
-      }
-      disconnect() {
-        callbacks.clear()
-      }
-    } as unknown as typeof globalThis.IntersectionObserver
-    triggerCenter = (el) => {
-      const cb = callbacks.get(el)
-      cb?.(
-        [{ target: el, isIntersecting: true } as IntersectionObserverEntry],
-        {} as IntersectionObserver,
-      )
-    }
-  })
-
-  afterEach(() => {
-    globalThis.IntersectionObserver = origIO
-    vi.useRealTimers()
-  })
-
-  const renderWithOnChange = () => {
-    const onChange = vi.fn()
-    const ref = createRef<CarouselHandlerRef>()
-    const { container } = render(
-      <Carousel navigationButtons indicator onChange={onChange} ref={ref}>
-        <div>0</div>
-        <div>1</div>
-        <div>2</div>
-      </Carousel>,
-    )
-    const scroller = container.querySelector(
-      '.charcoal-carousel__scroller',
-    ) as HTMLElement
-    // jsdom はレイアウトを持たず scrollWidth/clientWidth/scrollTo が無いため、
-    // nav ボタンの活性化とプログラム的スクロールが成立するようスタブする。
-    // 本物のブラウザ同様、scrollTo は scroll イベントを起こす（これが無いと
-    // IntersectionObserver の報告が移動より先に届く非現実的な順序になる）。
-    mockScrollerGeometry(scroller)
-    scroller.scrollTo = vi.fn(() => {
-      scroller.dispatchEvent(new Event('scroll'))
-    }) as unknown as typeof scroller.scrollTo
-    fireEvent.scroll(scroller)
-    const slides = container.querySelectorAll(
-      '.charcoal-carousel__scroller > *',
-    )
-    return { onChange, container, scroller, slides, ref }
-  }
-
-  // scrollend 非対応の jsdom では debounce(100ms) 経路になる
-  const settleScroll = (scroller: HTMLElement) => {
-    scroller.dispatchEvent(new Event('scroll'))
-    vi.advanceTimersByTime(150)
-  }
-
-  it('next ボタンの送りは source=navigation で発火する', () => {
-    vi.useFakeTimers()
-    try {
-      const { onChange, container, scroller, slides } = renderWithOnChange()
-      // マウント直後の静止（初期位置の適用）は帰属できないので発火しない
-      settleScroll(scroller)
-      expect(onChange).not.toHaveBeenCalled()
-
-      container
-        .querySelector<HTMLButtonElement>('[data-direction="next"]')
-        ?.click()
-      triggerCenter(slides[1])
-      settleScroll(scroller)
-
-      expect(onChange).toHaveBeenCalledExactlyOnceWith({
-        index: 1,
-        source: 'navigation',
-      })
-    } finally {
-      vi.useRealTimers()
-    }
-  })
-
-  it('← / → は source=keyboard で発火する', () => {
-    vi.useFakeTimers()
-    try {
-      const { onChange, scroller, slides } = renderWithOnChange()
-      fireEvent.keyDown(scroller, { key: 'ArrowRight' })
-      triggerCenter(slides[1])
-      settleScroll(scroller)
-
-      expect(onChange).toHaveBeenCalledExactlyOnceWith({
-        index: 1,
-        source: 'keyboard',
-      })
-    } finally {
-      vi.useRealTimers()
-    }
-  })
-
-  it('indicator の dot は source=indicator で発火する', () => {
-    vi.useFakeTimers()
-    // jsdom には scrollIntoView が無いのでモックを定義する。
-    Element.prototype.scrollIntoView = vi.fn()
-    try {
-      const { onChange, container, scroller, slides } = renderWithOnChange()
-      const dots = container.querySelectorAll(
-        '.charcoal-carousel__indicator__item',
-      )
-      act(() => {
-        fireEvent.click(dots[2])
-      })
-      // scrollIntoView のモックは scroll イベントを起こさないので明示的に起こす
-      fireEvent.scroll(scroller)
-      triggerCenter(slides[2])
-      settleScroll(scroller)
-
-      expect(onChange).toHaveBeenCalledExactlyOnceWith({
-        index: 2,
-        source: 'indicator',
-      })
-    } finally {
-      delete (Element.prototype as { scrollIntoView?: unknown }).scrollIntoView
-      vi.useRealTimers()
-    }
-  })
-
-  it('スワイプなどのポインタ操作は source=pointer で発火し、指を離すまで静止処理を保留する', () => {
-    vi.useFakeTimers()
-    try {
-      const { onChange, scroller, slides } = renderWithOnChange()
-      scroller.dispatchEvent(new Event('pointerdown', { bubbles: true }))
-      scroller.dispatchEvent(new Event('scroll'))
-      triggerCenter(slides[2])
-      // 指が触れたまま debounce が切れても発火しない
-      vi.advanceTimersByTime(150)
-      expect(onChange).not.toHaveBeenCalled()
-
-      window.dispatchEvent(new Event('pointerup', { bubbles: true }))
-      expect(onChange).toHaveBeenCalledExactlyOnceWith({
-        index: 2,
-        source: 'pointer',
-      })
-    } finally {
-      vi.useRealTimers()
-    }
-  })
-
-  it('タッチのパン開始で来る pointercancel では静止処理を保留したままにする', () => {
-    vi.useFakeTimers()
-    try {
-      const { onChange, scroller, slides } = renderWithOnChange()
-      // jsdom に PointerEvent は無いので pointerType だけを載せた Event で代用する
-      scroller.dispatchEvent(
-        Object.assign(new Event('pointerdown', { bubbles: true }), {
-          pointerType: 'touch',
-        }),
-      )
-      scroller.dispatchEvent(new Event('scroll'))
-      triggerCenter(slides[2])
-
-      // パンに転じた時点で pointercancel が来るが、指はまだ触れている
-      window.dispatchEvent(
-        Object.assign(new Event('pointercancel', { bubbles: true }), {
-          pointerType: 'touch',
-        }),
-      )
-      vi.advanceTimersByTime(150)
-      expect(onChange).not.toHaveBeenCalled()
-
-      window.dispatchEvent(new Event('touchend', { bubbles: true }))
-      expect(onChange).toHaveBeenCalledExactlyOnceWith({
-        index: 2,
-        source: 'pointer',
-      })
-    } finally {
-      vi.useRealTimers()
-    }
-  })
-
-  it('同じ index への着地では発火しない', () => {
-    vi.useFakeTimers()
-    try {
-      const { onChange, container, scroller, slides } = renderWithOnChange()
-      container
-        .querySelector<HTMLButtonElement>('[data-direction="next"]')
-        ?.click()
-      triggerCenter(slides[0])
-      settleScroll(scroller)
-
-      expect(onChange).not.toHaveBeenCalled()
-    } finally {
-      vi.useRealTimers()
-    }
-  })
-
-  it('resetScroll() による初期位置への復帰では発火せず、基準だけ進む', () => {
-    vi.useFakeTimers()
-    try {
-      const { onChange, container, scroller, slides, ref } =
-        renderWithOnChange()
-      container
-        .querySelector<HTMLButtonElement>('[data-direction="next"]')
-        ?.click()
-      triggerCenter(slides[1])
-      settleScroll(scroller)
-      expect(onChange).toHaveBeenCalledTimes(1)
-
-      act(() => ref.current?.resetScroll())
-      triggerCenter(slides[0])
-      settleScroll(scroller)
-      expect(onChange).toHaveBeenCalledTimes(1)
-
-      // 基準は 0 に戻っているので、0 への再着地は発火しない
-      container
-        .querySelector<HTMLButtonElement>('[data-direction="next"]')
-        ?.click()
-      triggerCenter(slides[0])
-      settleScroll(scroller)
-      expect(onChange).toHaveBeenCalledTimes(1)
-    } finally {
-      vi.useRealTimers()
-    }
-  })
-})
-
-describe('autoplay', () => {
-  let triggerCenter: (el: Element) => void
-  let origIO: typeof globalThis.IntersectionObserver
-
-  beforeEach(() => {
-    // jsdom は onscrollend を持つが実イベントは発火しないため debounce(100ms) 経路を強制する
-    Reflect.deleteProperty(window, 'onscrollend')
-    const callbacks = new Map<Element, IntersectionObserverCallback>()
-    origIO = globalThis.IntersectionObserver
-    globalThis.IntersectionObserver = class {
-      constructor(private cb: IntersectionObserverCallback) {}
-      observe(el: Element) {
-        callbacks.set(el, this.cb)
-      }
-      unobserve(el: Element) {
-        callbacks.delete(el)
-      }
-      disconnect() {
-        callbacks.clear()
-      }
-    } as unknown as typeof globalThis.IntersectionObserver
-    triggerCenter = (el) => {
-      callbacks.get(el)?.(
-        [{ target: el, isIntersecting: true } as IntersectionObserverEntry],
-        {} as IntersectionObserver,
-      )
-    }
-  })
-
-  afterEach(() => {
-    globalThis.IntersectionObserver = origIO
-    vi.useRealTimers()
-  })
-
-  // scrollLeft を反映し scroll イベントを発火するスタブ。静止（debounce settle）経路が
-  // 実際に走らないと滞留の数え直しを検証できないため、bare な vi.fn() では足りない。
-  const stubScrollTo = (scroller: HTMLElement) => {
-    let currentScrollLeft = 0
-    Object.defineProperty(scroller, 'scrollLeft', {
-      get: () => currentScrollLeft,
-      set: (v: number) => {
-        currentScrollLeft = v
-      },
-      configurable: true,
-    })
-    const scrollTo = vi.fn((opts: ScrollToOptions) => {
-      if (opts.left != null) currentScrollLeft = opts.left
-      scroller.dispatchEvent(new Event('scroll'))
-    })
-    scroller.scrollTo = scrollTo as unknown as typeof scroller.scrollTo
-    return scrollTo
-  }
-
-  const renderAutoplay = (
-    props: Partial<Pick<CarouselProps, 'autoplay' | 'onChange'>> = {},
-  ) => {
-    const { container } = render(
-      <Carousel autoplay={{ interval: 3000 }} {...props}>
-        <div>0</div>
-        <div>1</div>
-        <div>2</div>
-      </Carousel>,
-    )
-    const root = container.querySelector('.charcoal-carousel') as HTMLElement
-    const scroller = container.querySelector(
-      '.charcoal-carousel__scroller',
-    ) as HTMLElement
-    // jsdom はレイアウトを持たないので、送り先の算出に必要な寸法を与える
-    mockScrollerGeometry(scroller)
-    const scrollTo = stubScrollTo(scroller)
-    const slides = container.querySelectorAll(
-      '.charcoal-carousel__scroller > *',
-    )
-    return { container, root, scroller, scrollTo, slides }
-  }
-
-  const settle = () => vi.advanceTimersByTime(150)
-
-  it('interval 経過で次のスライドの静止位置へ smooth スクロールする', () => {
-    vi.useFakeTimers()
-    try {
-      const { scrollTo } = renderAutoplay()
-      expect(scrollTo).not.toHaveBeenCalled()
-      vi.advanceTimersByTime(3000)
-      // mockScrollerGeometry: clientWidth 800 / 子は offsetLeft i*400・offsetWidth 380。
-      // 既定は size M ＝ snapType 'none' なので start（実座標）に寄せる: 0, 400, 800。
-      expect(scrollTo).toHaveBeenCalledExactlyOnceWith({
-        left: 400,
-        behavior: 'smooth',
-      })
-    } finally {
-      vi.useRealTimers()
-    }
-  })
-
-  it('autoplay={true} は既定の 5000ms で回る', () => {
-    vi.useFakeTimers()
-    try {
-      const { scrollTo } = renderAutoplay({ autoplay: true })
-      vi.advanceTimersByTime(4999)
-      expect(scrollTo).not.toHaveBeenCalled()
-      vi.advanceTimersByTime(1)
-      expect(scrollTo).toHaveBeenCalledTimes(1)
-    } finally {
-      vi.useRealTimers()
-    }
-  })
-
-  it('autoplay 未指定 / false ならタイマーを張らない', () => {
-    vi.useFakeTimers()
-    try {
-      const a = renderAutoplay({ autoplay: undefined })
-      const b = renderAutoplay({ autoplay: false })
-      vi.advanceTimersByTime(60_000)
-      expect(a.scrollTo).not.toHaveBeenCalled()
-      expect(b.scrollTo).not.toHaveBeenCalled()
-    } finally {
-      vi.useRealTimers()
-    }
-  })
-
-  it.each([0, -1, Number.NaN, Number.POSITIVE_INFINITY, 2_147_483_648])(
-    '無効な interval=%s では既定値 (5000ms) にフォールバックし、開発時は警告する',
-    (interval) => {
-      vi.useFakeTimers()
-      const error = vi.spyOn(console, 'error').mockImplementation(() => {})
-      try {
-        const { scrollTo } = renderAutoplay({ autoplay: { interval } })
-        expect(error).toHaveBeenCalledWith(
-          expect.stringContaining('"interval"'),
-        )
-        vi.advanceTimersByTime(4999)
-        expect(scrollTo).not.toHaveBeenCalled()
-        vi.advanceTimersByTime(1)
-        expect(scrollTo).toHaveBeenCalledTimes(1)
-      } finally {
-        error.mockRestore()
-        vi.useRealTimers()
-      }
-    },
-  )
-
-  it('hover 中は進まず、外すと再開する', () => {
-    vi.useFakeTimers()
-    try {
-      const { root, scrollTo } = renderAutoplay()
-      act(() => {
-        root.dispatchEvent(
-          Object.assign(new Event('pointerenter'), { pointerType: 'mouse' }),
-        )
-      })
-      vi.advanceTimersByTime(3000)
-      expect(scrollTo).not.toHaveBeenCalled()
-
-      act(() => {
-        root.dispatchEvent(
-          Object.assign(new Event('pointerleave'), { pointerType: 'mouse' }),
-        )
-      })
-      vi.advanceTimersByTime(3000)
-      expect(scrollTo).toHaveBeenCalledTimes(1)
-    } finally {
-      vi.useRealTimers()
-    }
-  })
-
-  it('pauseOnHover: false なら hover でも止まらない', () => {
-    vi.useFakeTimers()
-    try {
-      const { root, scrollTo } = renderAutoplay({
-        autoplay: { interval: 3000, pauseOnHover: false },
-      })
-      act(() => {
-        root.dispatchEvent(
-          Object.assign(new Event('pointerenter'), { pointerType: 'mouse' }),
-        )
-      })
-      vi.advanceTimersByTime(3000)
-      expect(scrollTo).toHaveBeenCalledTimes(1)
-    } finally {
-      vi.useRealTimers()
-    }
-  })
-
-  it('マウント時に既にポインタが乗っていれば止まっている', () => {
-    vi.useFakeTimers()
-    stubbedHover = true
-    try {
-      const { scrollTo } = renderAutoplay()
-      vi.advanceTimersByTime(3000)
-      expect(scrollTo).not.toHaveBeenCalled()
-    } finally {
-      vi.useRealTimers()
-    }
-  })
-
-  it('キーボードフォーカスが Carousel 内にある間は進まない', () => {
-    vi.useFakeTimers()
-    try {
-      const { scroller, scrollTo } = renderAutoplay()
-      // react-aria は document 上の keydown でフォーカスの由来を keyboard と判定する
-      fireEvent.keyDown(document.body, { key: 'Tab' })
-      act(() => scroller.focus())
-      vi.advanceTimersByTime(3000)
-      expect(scrollTo).not.toHaveBeenCalled()
-
-      act(() => scroller.blur())
-      vi.advanceTimersByTime(3000)
-      expect(scrollTo).toHaveBeenCalledTimes(1)
-    } finally {
-      vi.useRealTimers()
-    }
-  })
-
-  it('手動スクロールの静止で滞留が頭から数え直される', () => {
-    vi.useFakeTimers()
-    try {
-      const { scroller, scrollTo } = renderAutoplay()
-      vi.advanceTimersByTime(2000)
-      scroller.dispatchEvent(new Event('scroll'))
-      settle()
-      // 到着（静止）は scroll(t=2000) から debounce(100ms) 後の t=2100。
-      // 滞留はそこから数え直すので次の送りは t=5100。
-      vi.advanceTimersByTime(2949)
-      expect(scrollTo).not.toHaveBeenCalled()
-      vi.advanceTimersByTime(1)
-      expect(scrollTo).toHaveBeenCalledTimes(1)
-    } finally {
-      vi.useRealTimers()
-    }
-  })
-
-  it('自動送りの滞留は到着（静止）から数える', () => {
-    vi.useFakeTimers()
-    try {
-      const { scrollTo } = renderAutoplay()
-      // t=3000: 送り。stubScrollTo が scroll を出し、debounce(100ms) で t=3100 に静止する。
-      vi.advanceTimersByTime(3000)
-      expect(scrollTo).toHaveBeenCalledTimes(1)
-      // t=6000: 送りから 3000ms 経ったが、到着（t=3100）からは 2900ms なので送らない
-      vi.advanceTimersByTime(3000)
-      expect(scrollTo).toHaveBeenCalledTimes(1)
-      // t=6100: 到着から 3000ms
-      vi.advanceTimersByTime(100)
-      expect(scrollTo).toHaveBeenCalledTimes(2)
-    } finally {
-      vi.useRealTimers()
-    }
-  })
-
-  it('自動送りの着地は onChange に source=auto で通知される', () => {
-    vi.useFakeTimers()
-    try {
-      const onChange = vi.fn()
-      const { slides } = renderAutoplay({ onChange })
-      vi.advanceTimersByTime(3000)
-      triggerCenter(slides[1])
-      settle()
-      expect(onChange).toHaveBeenCalledExactlyOnceWith({
-        index: 1,
-        source: 'auto',
-      })
-    } finally {
-      vi.useRealTimers()
-    }
-  })
-
-  it('自動送り中のスワイプは source=pointer に帰属する', () => {
-    vi.useFakeTimers()
-    try {
-      const onChange = vi.fn()
-      const { scroller, slides } = renderAutoplay({ onChange })
-      vi.advanceTimersByTime(3000)
-      scroller.dispatchEvent(new Event('pointerdown', { bubbles: true }))
-      scroller.dispatchEvent(new Event('scroll'))
-      window.dispatchEvent(new Event('pointerup', { bubbles: true }))
-      triggerCenter(slides[1])
-      settle()
-      expect(onChange).toHaveBeenCalledExactlyOnceWith({
-        index: 1,
-        source: 'pointer',
-      })
-    } finally {
-      vi.useRealTimers()
-    }
-  })
-
-  it('prefers-reduced-motion: reduce では一切進まない', () => {
-    vi.useFakeTimers()
-    const original = window.matchMedia
-    window.matchMedia = vi.fn(
-      () =>
-        ({
-          matches: true,
-          addEventListener: vi.fn(),
-          removeEventListener: vi.fn(),
-        }) as unknown as MediaQueryList,
-    ) as typeof window.matchMedia
-    try {
-      const { scrollTo } = renderAutoplay()
-      vi.advanceTimersByTime(60_000)
-      expect(scrollTo).not.toHaveBeenCalled()
-    } finally {
-      window.matchMedia = original
-      vi.useRealTimers()
-    }
-  })
-})
-
-describe('loop + onChange', () => {
-  // item の offsetLeft/offsetWidth は clone の作り直しに追従させるため、
-  // scroller 内の位置から導出する（describe('loop') と同じ手口）。
-  //
-  // slot 400 / clientWidth 800 では実 3 枚に対し clone が各端 9 枚（children 21）:
-  //   setWidth = children[12].offsetLeft − children[9].offsetLeft = 1200
-  //   maxScroll = 8400 − 800 = 7600 / 帯域 = 中央 [(7600−1200)/2, +1200) = [3200, 4400)
-  const SLOT_WIDTH = 400
-  const slotIndex = (el: HTMLElement) => {
-    const parent = el.parentElement
-    return parent?.classList.contains('charcoal-carousel__scroller')
-      ? Array.prototype.indexOf.call(parent.children, el)
-      : null
-  }
-  const originalOffsetLeft = Object.getOwnPropertyDescriptor(
-    HTMLElement.prototype,
-    'offsetLeft',
-  )
-  const originalOffsetWidth = Object.getOwnPropertyDescriptor(
-    HTMLElement.prototype,
-    'offsetWidth',
-  )
-
-  let roCallbacks: Array<(entries: unknown[]) => void>
-  let origRO: typeof globalThis.ResizeObserver
-  let origIO: typeof globalThis.IntersectionObserver
-  let triggerCenter: (el: Element) => void
-
-  beforeEach(() => {
-    // jsdom は onscrollend を持つが実イベントは発火しないため debounce(100ms) 経路を強制する
-    Reflect.deleteProperty(window, 'onscrollend')
-    Object.defineProperty(HTMLElement.prototype, 'offsetLeft', {
-      configurable: true,
-      get(this: HTMLElement) {
-        const i = slotIndex(this)
-        return i == null ? 0 : i * SLOT_WIDTH
-      },
-    })
-    Object.defineProperty(HTMLElement.prototype, 'offsetWidth', {
-      configurable: true,
-      get(this: HTMLElement) {
-        return slotIndex(this) == null ? 0 : SLOT_WIDTH - 20
-      },
-    })
-
-    roCallbacks = []
-    origRO = globalThis.ResizeObserver
-    globalThis.ResizeObserver = class {
-      observe = vi.fn()
-      unobserve = vi.fn()
-      disconnect = vi.fn()
-      constructor(cb: (entries: unknown[]) => void) {
-        roCallbacks.push(cb)
-      }
-    } as unknown as typeof globalThis.ResizeObserver
-
-    const callbacks = new Map<Element, IntersectionObserverCallback>()
-    origIO = globalThis.IntersectionObserver
-    globalThis.IntersectionObserver = class {
-      constructor(private cb: IntersectionObserverCallback) {}
-      observe(el: Element) {
-        callbacks.set(el, this.cb)
-      }
-      unobserve(el: Element) {
-        callbacks.delete(el)
-      }
-      disconnect() {
-        callbacks.clear()
-      }
-    } as unknown as typeof globalThis.IntersectionObserver
-    triggerCenter = (el) => {
-      callbacks.get(el)?.(
-        [{ target: el, isIntersecting: true } as IntersectionObserverEntry],
-        {} as IntersectionObserver,
-      )
-    }
-  })
-
-  afterEach(() => {
-    if (originalOffsetLeft) {
-      Object.defineProperty(
-        HTMLElement.prototype,
-        'offsetLeft',
-        originalOffsetLeft,
-      )
-    }
-    if (originalOffsetWidth) {
-      Object.defineProperty(
-        HTMLElement.prototype,
-        'offsetWidth',
-        originalOffsetWidth,
-      )
-    }
-    globalThis.ResizeObserver = origRO
-    globalThis.IntersectionObserver = origIO
-    vi.useRealTimers()
-  })
-
-  it('clone 帯への自動送りは、静止後のテレポートを挟んでも source=auto で 1 回だけ発火する', () => {
-    vi.useFakeTimers()
-    try {
-      const onChange = vi.fn()
-      render(
-        <Carousel loop autoplay={{ interval: 3000 }} onChange={onChange}>
-          <div>0</div>
-          <div>1</div>
-          <div>2</div>
-        </Carousel>,
-      )
-      const scroller = getScroller()
-      let scrollLeft = 0
-      Object.defineProperty(scroller, 'scrollLeft', {
-        get: () => scrollLeft,
-        set: (v: number) => {
-          scrollLeft = v
-        },
-        configurable: true,
-      })
-      Object.defineProperty(scroller, 'scrollWidth', {
-        value: 8400,
-        configurable: true,
-      })
-      Object.defineProperty(scroller, 'clientWidth', {
-        value: 800,
-        configurable: true,
-      })
-      // 本物のブラウザ同様、scrollTo は scroll イベントを起こす
-      const scrollTo = vi.fn((opts: ScrollToOptions) => {
-        if (opts.left != null) scrollLeft = opts.left
-        scroller.dispatchEvent(new Event('scroll'))
-      })
-      scroller.scrollTo = scrollTo as unknown as typeof scroller.scrollTo
-
-      // 幾何の実測 → clone 枚数の確定 → 初期位置（実セット先頭 = 3600）の適用
-      act(() => {
-        roCallbacks.forEach((cb) => cb([{ target: scroller }]))
-      })
-      expect(scroller.children).toHaveLength(21)
-      act(() => {
-        vi.advanceTimersByTime(150)
-      })
-      expect(scrollLeft).toBe(3600)
-
-      // 帯域上限（4400）の直前まで手で送っておき、次の自動送りで帯域外へ出す
-      act(() => {
-        scrollLeft = 4400
-        scroller.dispatchEvent(new Event('scroll'))
-        vi.advanceTimersByTime(150)
-      })
-      expect(onChange).not.toHaveBeenCalled()
-      scrollTo.mockClear()
-
-      // t+3000: 自動送りが clone-after 帯（children[12] = 4800）へ進む
-      act(() => {
-        vi.advanceTimersByTime(3000)
-      })
-      expect(scrollTo).toHaveBeenCalledExactlyOnceWith({
-        left: 4800,
-        behavior: 'smooth',
-      })
-
-      // 静止で 4800 は帯域外と判明し、合同位置 3600 へテレポートする
-      act(() => {
-        vi.advanceTimersByTime(150)
-      })
-      expect(scrollTo).toHaveBeenLastCalledWith({
-        left: 3600,
-        behavior: 'instant',
-      })
-      expect(onChange).not.toHaveBeenCalled()
-
-      // 中央検出はテレポートの飛行中に遅れて届く（clone-after の 2 枚目＝論理 index 1）
-      act(() => {
-        triggerCenter(scroller.children[13])
-      })
-      act(() => {
-        vi.advanceTimersByTime(150)
-      })
-
-      expect(onChange).toHaveBeenCalledExactlyOnceWith({
-        index: 1,
-        source: 'auto',
-      })
-    } finally {
-      vi.useRealTimers()
-    }
   })
 })
