+++ packages/react/src/components/Carousel/intent/scrollIntent.test.ts
